+++ test/tracing/httpClient/clientApp.js
@@ -13,13 +13,18 @@ require('../../../')({
 var bodyParser = require('body-parser');
 var rp = require('request-promise');
 var express = require('express');
+var semver = require('semver');
 var morgan = require('morgan');
 var AWS = require('aws-sdk');
 var path = require('path');
 var fs = require('fs');
 
+// WHATWG URL class is globally availabe as of Node.js 10.0.0, needs to be required in older versions.
+var URL = semver.lt(process.versions.node, '10.0.0') ? require('url').URL : global.URL;
+
 var httpModule = process.env.USE_HTTPS === 'true' ? require('https') : require('http');
 var protocol = process.env.USE_HTTPS === 'true' ? 'https' : 'http';
+var baseUrl = protocol + '://127.0.0.1:' + process.env.SERVER_PORT;
 
 var app = express();
 var awsRegion = process.env.AWS_REGION || 'eu-west-1';
@@ -35,7 +40,7 @@ app.use(bodyParser.json());
 app.get('/', function(req, res) {
   rp({
     method: 'GET',
-    url: protocol + '://127.0.0.1:' + process.env.SERVER_PORT + '/',
+    url: baseUrl + '/',
     strictSSL: false
   })
     .then(function() {
@@ -48,7 +53,7 @@ app.get('/', function(req, res) {
 
 app.get('/request-url-and-options', function(req, res) {
   httpModule.request(
-    protocol + '://127.0.0.1:' + process.env.SERVER_PORT + '/request-url-opts',
+    createUrl(req, '/request-url-opts'),
     { rejectUnauthorized: false },
     function() {
       return res.sendStatus(200);
@@ -58,7 +63,7 @@ app.get('/request-url-and-options', function(req, res) {
 
 app.get('/request-url-only', function(req, res) {
   httpModule.request(
-    protocol + '://127.0.0.1:' + process.env.SERVER_PORT + '/request-url',
+    createUrl(req, '/request-only-url'),
     function() {
       return res.sendStatus(200);
     }
@@ -70,7 +75,7 @@ app.get('/request-options-only', function(req, res) {
     hostname: '127.0.0.1',
     port: process.env.SERVER_PORT,
     method: 'GET',
-    path: '/request-opts',
+    path: '/request-only-opts',
     rejectUnauthorized: false,
   }, function() {
     return res.sendStatus(200);
@@ -79,7 +84,7 @@ app.get('/request-options-only', function(req, res) {
 
 app.get('/get-url-and-options', function(req, res) {
   httpModule.get(
-    protocol + '://127.0.0.1:' + process.env.SERVER_PORT + '/get-url-opts',
+    createUrl(req, '/get-url-opts'),
     { rejectUnauthorized: false },
     function() {
       return res.sendStatus(200);
@@ -89,7 +94,7 @@ app.get('/get-url-and-options', function(req, res) {
 
 app.get('/get-url-only', function(req, res) {
   httpModule.get(
-    protocol + '://127.0.0.1:' + process.env.SERVER_PORT + '/get-url',
+    createUrl(req, '/get-only-url'),
     function() {
       return res.sendStatus(200);
     }
@@ -101,7 +106,7 @@ app.get('/get-options-only', function(req, res) {
     hostname: '127.0.0.1',
     port: process.env.SERVER_PORT,
     method: 'GET',
-    path: '/get-opts',
+    path: '/get-only-opts',
     rejectUnauthorized: false,
   }, function() {
     return res.sendStatus(200);
@@ -111,7 +116,7 @@ app.get('/get-options-only', function(req, res) {
 app.get('/timeout', function(req, res) {
   rp({
     method: 'GET',
-    url: protocol + '://127.0.0.1:' + process.env.SERVER_PORT + '/timeout',
+    url: baseUrl + '/timeout',
     timeout: 500,
     strictSSL: false
   })
@@ -210,6 +215,12 @@ if (process.env.USE_HTTPS === 'true') {
   app.listen(process.env.APP_PORT, function() {
     log('Listening on port: ' + process.env.APP_PORT);
   });
+}
+
+function createUrl(req, urlPath) {
+  return req.query.urlObject ?
+    new URL(urlPath, baseUrl) :
+    baseUrl + urlPath;
 }
 
 function log() {